+++ changelog.md
@@ -1,7 +1,22 @@
-31 May 2020
+# 31 May 2020
+## 0.3
+### ADDED
+  1. ModalCollection, ToasterCollection
+  2. Validation input file PDF and signature
+  3. loading state while loading page PDF and download
+  4. Dynamic margin-bottom on card-body page for PDF refer to PreviewControl Height
+  5. End Modal with restart function with vuex
+### CHANGED
+  1. Reformating Preview Control for PreviewHeaderControlTarget, PreviewHeaderControlDownload
+### DELETED 
+  1. Unnecessary Commented
+
+## 0.2
 ### CHANGED
   1. Change FrontEnd logic to vuex
-30 May 2020
+
+# 30 May 2020
+## 0.1
 ### ADDED
   1. Create FrontEnd (App, Components, Preview, SignerPad, Stepper,CustomStyle)
   2. Create BackEnd (SignerController)